Refactor: DashboardRepository으로 네이밍 변경
